moved google auth to auth-router
removed all renders to hbs file from google auth. changed views to ejs
