--- api/views/layout.hbs
@@ -1,11 +1,0 @@
-<!doctype html>
-<html>
-	<head>
-		<meta charset="utf-8">
-		<title>Google OAuth2.0 Template</title>
-		<link rel='stylesheet' type='text/css' href='style.css'>
-	</head>
-	<body>
-		{{{ body }}}
-	</body>
-</html>
