--- api/views/login.hbs
@@ -1,1 +1,0 @@
-<a href="/auth/google" class='button'>Login with Google</a>
