+++ views/index.ejs
@@ -1,0 +1,1 @@
+<h1>Hi</h1>
