+++ views/login.ejs
@@ -1,0 +1,16 @@
+<h1>Login</h1>
+
+<% if (messages.error) { %> <%= messages.error %> <% } %>
+
+<form action="/login" method="POST">
+	<div>
+		<label for="username">Username</label>
+		<input type="username" id="username" name="username" required />
+	</div>
+	<div>
+		<label for="password">Password</label>
+		<input type="password" id="password" name="password" required />
+	</div>
+	<button type="submit">Login</button>
+</form>
+<a href="/register">Register</a>
