+++ views/register.ejs
@@ -1,0 +1,24 @@
+<h1>Register</h1>
+
+<% if (messages.error) { %> <%= messages.error %> <% } %>
+
+<form action="/register" method="POST">
+	<div>
+		<label for="name">Name</label>
+		<input type="name" id="name" name="name" required />
+	</div>
+	<div>
+		<label for="username">Username</label>
+		<input type="username" id="username" name="username" required />
+	</div>
+	<div>
+		<label for="password">Password</label>
+		<input type="password" id="password" name="password" required />
+	</div>
+	<div>
+		<label for="email">Email</label>
+		<input type="email" id="email" name="email" />
+	</div>
+	<button type="submit">Register</button>
+</form>
+<a href="/login">Login</a>
